pyproject.toml
[tool.poetry]
name = "flitzserver"
version = "0.1.0"
description = ""
authors = ["Gyuhwan Park★ <unstabler@unstabler.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
uuid-v7 = "^1.0.0"
djangorestframework = "^3.15.2"
pyjwt = "^2.10.1"
dacite = "^1.8.1"
django-storages = {extras = ["s3"], version = "^1.14.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
